#240: rename some files in tests/
