+++ tests/command_generate_scanner.py
